rename ncUtils
