+++ util/ncUtils.js
